+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added localeflags/contentflags to more info modal for files.
 - Added 52 new names to the contenthash namer.
 - Added available/unavailable filter keywords to files page (only useful with showAllFiles setting on).
+- Added encryptedbutnot,lookupmatch and lookupwrong filter keywords to files page.
 - Prefer spawning a new TACTSharp instance over using wago.tools during diffs.
 - Fixed missing encryption icons in diffs.
 - Fixed contenthashes not resetting between build loads.